requirements.txt
nose==1.3.3
PyJWT==0.2.1
